+++ tests/testthat/test-find-netlogo-console.R
@@ -1,0 +1,7 @@
+testthat::test_that("`find_netlogo_console()` | General test", {
+
+})
+
+testthat::test_that("`find_netlogo_console()` | Error test", {
+
+})