+++ tests/testthat/test-find-netlogo-home.R
@@ -1,0 +1,7 @@
+testthat::test_that("`find_netlogo_home()` | General test", {
+
+})
+
+testthat::test_that("`find_netlogo_home()` | Error test", {
+
+})