+++ tests/testthat/test-find-netlogo-version.R
@@ -1,0 +1,7 @@
+testthat::test_that("`find_netlogo_version()` | General test", {
+
+})
+
+testthat::test_that("`find_netlogo_version()` | Error test", {
+
+})